# Optimize content search