# Problem

The original `find.sh` script had several inefficiencies:

1. **Multiple temporary files**: Used 4+ temp files for intermediate processing
2. **O(n) jq invocations**: Called jq once per directory and once per matching file
3. **Complex pipeline**: Multiple file reads and writes in loops

# Solution

#### 1. Eliminated Multiple Temporary Files
**Before**: Used 4+ temp files (`TMP_FILES`, `TMP_RESULTS`, `TMP_RESULTS.dirlist`, `TMP_RESULTS.result`)
**After**: Single pipeline with no intermediate files (i.e. minimal I/O)

#### 2. Reduced jq Invocations from O(n) to O(1)
**Before**:
- 1 initial jq call to create structure
- 1 jq call per directory (to add to directories array)
- 1 jq call per matching file (could be hundreds)
- 1 final jq call to sort

**After**: Single jq operation produces `folders` result, interpolates the end result directly.

#### 3. Streamlined Data Processing Pipeline
**Before**: `rg -> rg -> sed -> sort -> while read -> jq (per file)`
**After**: `rg -> rg -> jq`

# Benchmarks

* Searching for increasingly narrow search terms yields results in just milliseconds on average.
* Searched "e" on an English language rom set (13k raw hits) to test worst case performance.

```
Term   Script         Min Time    Max Time   Avg Time   Median Time  Performance
--------------------------------------------------------------------------------------------
zelda  find-old.sh    0.1816s     0.2134s    0.1947s    0.1953s          
       find-new.sh    0.0305s     0.0342s    0.0321s    0.0317s           
       Improvement:                                                  +83.00% (6.0x faster)
--------------------------------------------------------------------------------------------
tennis find-old.sh    0.5354s     0.5532s    0.5435s    0.5392s         
       find-new.sh    0.0315s     0.0362s    0.0336s    0.0335s              
       Improvement:                                                  +93.00% (16.1x faster)
--------------------------------------------------------------------------------------------
the    find-old.sh    13.2671s    18.9075s   15.6527s   15.3009s              
       find-new.sh    0.0518s     0.0618s    0.0542s    0.0523s              
       Improvement:                                                  +99.00% (288.7x faster)
--------------------------------------------------------------------------------------------
e      find-old.sh    157.4212s   168.5029s  162.9513s  163.5174s            
       find-new.sh    0.1599s     0.3567s    0.2217s    0.1967s          
       Improvement:                                                  +99.00% (735.0x faster)
--------------------------------------------------------------------------------------------
```

# Notes 

* Note this was performed on Apple M3 Pro, searching 13k+ synthetic rom files
* On an RG35xx Plus performance is > 2x slower than these numbers 🙃