@xonglebongle 2025-07-20
Upon further testing, this implementation suffers from IO blocking issues.  The script fails with complex content structures, including multi-tiered directories.  Changing directories and running the script results in, seemingly random occurrences of, an unkillable zombie process.  Switching between "Local" and "Global" search across directories consistently causes system lock-ups.