@CatalyticArkun 2025-07-20
It's not that difficult to lock up the system - intense process calls can easily flood the system and cause stutters or freezes.