@gleepglopdev 2025-07-23
@xonglebongle I wasn't able to reproduce any lock-ups, but I'm happy to continue trying with any steps you can provide.

I also wonder what version of MuOS you experienced that issue on, apart from the potential differences in the environment, as well as the ROM files structure you mentioned above.

I'm working with the latest fresh install of MuOS Pixie, on an RG35xx Plus.