@xonglebongle 2025-07-25
Apologies for the late replies, has been a busy week.  This has lots of potential I can see that already.  We are coming closer to the Goose release so this may have to stay open until such time.