@gleepglopdev 2025-08-30
Congrats on launching Goose! 🐦 

I was still not able to reproduce the issues described above with this script on `MustardOS_RG40XX-H_2508.0_GOOSE-f9f63c91`.