@gleepglopdev 2025-11-09
@xonglebongle This is ready for another review. 

* Assuming the crashes you experienced weren't to do with my first commit's approach with a single `rg | rg | jq` pipeline and without temp files for search I've restored that approach as it's running far faster on my RG35XX Plus and RG 40XXV.
  * Recall before it was: rg -> rg -> sed -> sort -> while read -> jq (per matching file)
* I've also restored generating friendly name JSON now that you've updated the expected file path.
* Global and local search are working well for me, as is navigating out and back to content search (to re-render a previously generated search.json).